+++ CHANGELOG.md
@@ -22,6 +22,7 @@ CHANGELOG for FlatCAM beta
 - in Extract Drills Plugin updated the GUI
 - in Paint Plugin updated the GUI
 - in NCC Plugin updated the GUI
+- in Subtraction Plugin updated the GUI
 
 4.09.2021